+++ extension/src/popup/components/mnemonicPhrase/DisplayMnemonicPhrase/styles.scss
@@ -1,0 +1,73 @@
+.DisplayMnemonicPhrase {
+  &--header {
+    font-size: 2.5rem;
+    line-height: 3rem;
+    font-weight: var(--font-weight-normal);
+    color: var(--pal-text-primary);
+    text-align: left;
+  }
+
+  &--content {
+    margin-top: 1.5rem;
+    font-size: 1rem;
+    line-height: 1.5rem;
+    font-weight: var(--font-weight-normal);
+    color: var(--pal-text-tertiary);
+  }
+
+  &--screen {
+    display: flex;
+    flex-flow: row wrap;
+    align-content: center;
+    justify-content: space-between;
+    // TODO - use HEADER_HEIGHT constant
+    height: calc(100vh - 119px);
+    max-height: 40rem;
+    max-width: 57rem;
+    width: 100%;
+    margin: auto;
+  }
+
+  &--half-screen {
+    display: flex;
+    flex-direction: column;
+    justify-content: start;
+    padding: 2rem 0 2rem 1.55rem;
+    width: 27rem;
+  }
+
+  &--mnemonic-display {
+    margin: 2rem 0 1rem;
+    position: relative;
+    text-align: center;
+  }
+
+  &--ordered-list {
+    columns: 2;
+    text-align: start;
+    counter-reset: item;
+    list-style-type: none;
+
+    li {
+      padding-left: 1rem;
+    }
+
+    li::before {
+      color: var(--pal-text-tertiary);
+      content: none;
+      counter-increment: item;
+      content: counter(item);
+    }
+  }
+
+  &--display-buttons {
+    margin-bottom: 2.5rem;
+    margin-right: 1rem;
+    position: relative;
+    text-align: right;
+
+    img {
+      margin-left: 0.5rem;
+    }
+  }
+}
